requirements.txt
africastalking==1.2.4
aiofiles==22.1.0
aiosqlite==0.18.0
annotated-types==0.7.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
autopep8==1.7.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bokeh==3.2.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.3
colorcet==3.0.1
comm==0.1.3
configobj==5.0.8
configparser==7.0.0
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.8
distro==1.8.0
dnspython==2.6.1
email_validator==2.2.0
etelemetry==0.3.1
executing==1.2.0
fastapi-cli==0.0.4
fastjsonschema==2.16.3
filelock==3.15.4
fitz==0.0.1.dev2
Flask==2.2.2
fonttools==4.40.0
fpdf==1.7.2
fqdn==1.5.1
gotrue==2.5.4
h11==0.14.0
holoviews==1.17.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
hvplot==0.8.4
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.0
kiwisolver==1.4.4
linkify-it-py==2.0.2
looseversion==1.3.0
lxml==5.2.2
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==2.0.5
mysql-connector-python==8.0.32
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
orjson==3.10.5
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
panel==1.2.3
param==1.13.0
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2024.0.1
platformdirs==4.2.2
postgrest==0.16.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
prov==2.0.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pyct==0.5.0
pydantic==2.7.4
pydantic-extra-types==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.18.4
pydot==3.0.1
Pygments==2.15.0
pylightxl==1.60
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.0.9
pypdf==4.2.0
pypng==0.20220715.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytube==12.1.0
pytz==2023.3
pyviz_comms==3.0.0
pyxnat==1.6.2
PyYAML==6.0
pyzmq==25.0.2
qrcode==7.4.2
qtconsole==5.4.2
QtPy==2.3.1
rdflib==6.3.2
realtime==1.0.6
requests==2.32.3
resend==2.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
schema==0.7.5
scipy==1.14.0
Send2Trash==1.8.0
shellingham==1.5.4
shortuuid==1.0.9
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.37.2
storage3==0.7.6
StrEnum==0.4.15
supafunc==0.4.6
tabula==1.0.5
tabula-py==2.9.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
traits==6.3.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
uc-micro-py==1.0.2
ujson==5.10.0
uri-template==1.2.0
urllib3==1.26.12
uuid==1.30
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.22.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==12.0
Werkzeug==2.2.2
widgetsnbextension==4.0.7
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
xyzservices==2023.10.0
y-py==0.5.9
ypy-websocket==0.8.2
